+++ src/aktualizr_secondary/update_agent_file.h
@@ -13,6 +13,7 @@ class FileUpdateAgent : public UpdateAgent {
   bool getInstalledImageInfo(Uptane::InstalledImageInfo& installed_image_info) const override;
   bool download(const Uptane::Target& target, const std::string& data) override;
   data::ResultCode::Numeric install(const Uptane::Target& target) override;
+  void completeInstall() override;
   data::InstallationResult applyPendingInstall(const Uptane::Target& target) override;
 
  private: